Add files via upload
Face detection is a computer technology used to locate human faces in digital images or videos. It is a subfield of computer vision and pattern recognition. The goal of face detection is to identify and locate the faces in an image or video, often as a pre-processing step for other applications such as face recognition, emotion recognition, or facial expression analysis.